Added: bin/workflow for workflow generation
