+++ templates/workflows/release/VERSION
@@ -1,0 +1,1 @@
+1.0.0
