+++ test/release/js/.github/workflows/release.yml
@@ -1,0 +1,154 @@
+# GENERATED by sdk-actions `bin/workflow` — update with: bin/workflow update <this file>
+# sdk-actions: {"template":"release/js/turnkey","ref":"38dcec863c910f6802c03a94ab9aeb79931a0a3a","version":"1.0.0","params":{"access":"public","allowed_channels":"latest,rc,next,beta","dry_run_environment":"publish-dry-run","emoji":":javascript:","node_version":".nvmrc","package_name":"@braintrust/bt-publishing-test","provenance":"true","publish_environment":"publish","slack_channel_var":"SLACK_SDK_RELEASE_CHANNEL","slack_token_secret":"SLACK_BOT_TOKEN","workflow_name":"Release","working_directory":"."}}
+
+name: Release
+
+# Releases @braintrust/bt-publishing-test to npm via braintrustdata/sdk-actions. Dispatch after merging
+# the commit that bumps package.json's "version"; the release runs against that commit's SHA.
+on:
+  workflow_dispatch:
+    inputs:
+      release_type:
+        description: "Release type (stable → GitHub release; prerelease → none)"
+        type: choice
+        default: stable
+        options: [stable, prerelease]
+      channel:
+        description: "npm dist-tag / channel"
+        type: choice
+        default: latest
+        options: [latest, rc, next, beta]
+      sha:
+        description: "Commit SHA (of the version bump) to release"
+        required: true
+        type: string
+      prev_release:
+        description: "Release-notes anchor: a tag or SHA (empty → the previous tag)"
+        type: string
+        required: false
+      dry_run:
+        description: "Dry run: build without tagging or publishing"
+        type: boolean
+        default: false
+
+jobs:
+  configure:
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions:
+      contents: read
+    outputs:
+      version: ${{ steps.configure.outputs.version }}
+      release_tag: ${{ steps.configure.outputs.release_tag }}
+      prev_release: ${{ steps.configure.outputs.prev_release }}
+      branch: ${{ steps.configure.outputs.branch }}
+      on_release_branch: ${{ steps.configure.outputs.on_release_branch }}
+      commit_message: ${{ steps.configure.outputs.commit_message }}
+      channel: ${{ steps.configure.outputs.channel }}
+      github_release: ${{ steps.configure.outputs.github_release }}
+    steps:
+      - name: Configure release
+        id: configure
+        uses: braintrustdata/sdk-actions/actions/release/lang/js/configure@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          sha: ${{ inputs.sha }}
+          working_directory: .
+          release_type: ${{ inputs.release_type }}
+          channel: ${{ inputs.channel }}
+
+  prepare:
+    needs: configure
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions:
+      contents: write  # releases/generate-notes API
+    outputs:
+      pr_list: ${{ steps.prepare.outputs.pr_list }}
+      notes: ${{ steps.prepare.outputs.notes }}
+    steps:
+      - name: Prepare release
+        id: prepare
+        uses: braintrustdata/sdk-actions/actions/release/prepare@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          sha: ${{ inputs.sha }}
+          prev_release: ${{ inputs.prev_release || needs.configure.outputs.prev_release }}
+
+  validate:
+    needs: [configure, prepare]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 10
+    permissions:
+      contents: read
+    steps:
+      - name: Validate release
+        uses: braintrustdata/sdk-actions/actions/release/lang/js/validate@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          version: ${{ needs.configure.outputs.version }}
+          sha: ${{ inputs.sha }}
+          dry_run: ${{ inputs.dry_run }}
+          working_directory: .
+          node_version: .nvmrc
+          package_name: '@braintrust/bt-publishing-test'
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          channel: ${{ needs.configure.outputs.channel }}
+          allowed_channels: 'latest,rc,next,beta'
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          notes: ${{ needs.prepare.outputs.notes }}
+
+  request-approval:
+    needs: [configure, prepare, validate]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions: {}
+    steps:
+      - name: Request release approval
+        uses: braintrustdata/sdk-actions/actions/release/request-approval@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          sha: ${{ inputs.sha }}
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          prev_release: ${{ needs.configure.outputs.prev_release }}
+          branch: ${{ needs.configure.outputs.branch }}
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          commit_message: ${{ needs.configure.outputs.commit_message }}
+          pr_list: ${{ needs.prepare.outputs.pr_list }}
+          notes: ${{ needs.prepare.outputs.notes }}
+          dry_run: ${{ inputs.dry_run }}
+          slack_token: ${{ secrets.SLACK_BOT_TOKEN }}
+          slack_channel: ${{ vars.SLACK_SDK_RELEASE_CHANNEL }}
+          label: '@braintrust/bt-publishing-test'
+          emoji: ':javascript:'
+
+  build-and-ship:
+    needs: [configure, prepare, validate, request-approval]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 15
+    environment: ${{ inputs.dry_run && 'publish-dry-run' || 'publish' }}
+    permissions:
+      contents: write
+      id-token: write  # OIDC trusted publishing + provenance
+      attestations: write  # signed SBOM attestation
+    steps:
+      - name: Build and ship
+        uses: braintrustdata/sdk-actions/actions/release/lang/js/build-and-ship@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          sha: ${{ inputs.sha }}
+          working_directory: .
+          node_version: .nvmrc
+          dry_run: ${{ inputs.dry_run }}
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          github_release: ${{ needs.configure.outputs.github_release }}
+          version: ${{ needs.configure.outputs.version }}
+          access: 'public'
+          provenance: 'true'
+          channel: ${{ needs.configure.outputs.channel }}
+          package_name: '@braintrust/bt-publishing-test'
+          label: '@braintrust/bt-publishing-test'
+          notes: ${{ needs.prepare.outputs.notes }}
+          prev_release: ${{ needs.configure.outputs.prev_release }}
+          branch: ${{ needs.configure.outputs.branch }}
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          pr_list: ${{ needs.prepare.outputs.pr_list }}
+          slack_token: ${{ secrets.SLACK_BOT_TOKEN }}
+          slack_channel: ${{ vars.SLACK_SDK_RELEASE_CHANNEL }}
+          emoji: ':javascript:'
