+++ test/release/py/.github/workflows/release-custom.yml
@@ -1,0 +1,185 @@
+# GENERATED by sdk-actions `bin/workflow` — update with: bin/workflow update <this file>
+# sdk-actions: {"template":"release/py/custom","ref":"38dcec863c910f6802c03a94ab9aeb79931a0a3a","version":"1.0.0","params":{"dry_run_environment":"publish-dry-run","emoji":":python:","package_name":"bt-publishing-test","publish_environment":"publish","python_version":".python-version","slack_channel_var":"SLACK_SDK_RELEASE_CHANNEL","slack_token_secret":"SLACK_BOT_TOKEN","version_file":"","workflow_name":"Release","working_directory":"."}}
+
+name: Release
+
+# Releases bt-publishing-test to PyPI via braintrustdata/sdk-actions, custom-build shape:
+# an unprivileged `pack` job builds + attests dist/ (wheel + sdist) with no publish creds
+# and uploads it; a gated `ship-package` job verifies the attestation and publishes THOSE
+# exact dists. Dispatch after merging the commit that bumps VERSION in src/bt_publishing_test/version.py.
+on:
+  workflow_dispatch:
+    inputs:
+      release_type:
+        description: "Release type (stable → GitHub release; prerelease → none)"
+        type: choice
+        default: stable
+        options: [stable, prerelease]
+      sha:
+        description: "Commit SHA (of the version bump) to release"
+        required: true
+        type: string
+      prev_release:
+        description: "Release-notes anchor: a tag or SHA (empty → the previous tag)"
+        type: string
+        required: false
+      dry_run:
+        description: "Dry run: build + pack without attesting or publishing"
+        type: boolean
+        default: false
+
+jobs:
+  configure:
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions:
+      contents: read
+    outputs:
+      version: ${{ steps.configure.outputs.version }}
+      release_tag: ${{ steps.configure.outputs.release_tag }}
+      prev_release: ${{ steps.configure.outputs.prev_release }}
+      branch: ${{ steps.configure.outputs.branch }}
+      on_release_branch: ${{ steps.configure.outputs.on_release_branch }}
+      commit_message: ${{ steps.configure.outputs.commit_message }}
+      github_release: ${{ steps.configure.outputs.github_release }}
+    steps:
+      - name: Configure release
+        id: configure
+        uses: braintrustdata/sdk-actions/actions/release/lang/py/configure@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          version_file: src/bt_publishing_test/version.py
+          sha: ${{ inputs.sha }}
+          working_directory: .
+          release_type: ${{ inputs.release_type }}
+
+  prepare:
+    needs: configure
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions:
+      contents: write  # releases/generate-notes API
+    outputs:
+      pr_list: ${{ steps.prepare.outputs.pr_list }}
+      notes: ${{ steps.prepare.outputs.notes }}
+    steps:
+      - name: Prepare release
+        id: prepare
+        uses: braintrustdata/sdk-actions/actions/release/prepare@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          sha: ${{ inputs.sha }}
+          prev_release: ${{ inputs.prev_release || needs.configure.outputs.prev_release }}
+
+  validate:
+    needs: [configure, prepare]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 10
+    permissions:
+      contents: read
+    steps:
+      - name: Validate release
+        uses: braintrustdata/sdk-actions/actions/release/lang/py/validate@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          version: ${{ needs.configure.outputs.version }}
+          sha: ${{ inputs.sha }}
+          dry_run: ${{ inputs.dry_run }}
+          working_directory: .
+          python_version: .python-version
+          package_name: bt-publishing-test
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          notes: ${{ needs.prepare.outputs.notes }}
+          build: 'false'   # the pack job owns the build
+          sbom: 'false'    # the pack job generates + attests the SBOM
+
+  # BUILD side — unprivileged: builds + attests dist/, NO publish credentials.
+  pack:
+    needs: [configure, prepare, validate]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 15
+    permissions:
+      contents: read
+      id-token: write
+      attestations: write
+    steps:
+      - name: Pack + attest
+        uses: braintrustdata/sdk-actions/actions/release/lang/py/pack@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          sha: ${{ inputs.sha }}
+          working_directory: .
+          python_version: .python-version
+          dry_run: ${{ inputs.dry_run }}
+
+      - name: Stage artifact
+        run: |
+          mkdir -p artifact/dist
+          cp ./dist/*.whl ./dist/*.tar.gz artifact/dist/
+          cp ./sbom.json artifact/ 2>/dev/null || true
+
+      - name: Upload artifact
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
+        with:
+          name: release-package
+          path: artifact/
+          retention-days: 1
+          if-no-files-found: error
+
+  request-approval:
+    needs: [configure, prepare, pack]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions: {}
+    steps:
+      - name: Request release approval
+        uses: braintrustdata/sdk-actions/actions/release/request-approval@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          sha: ${{ inputs.sha }}
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          prev_release: ${{ needs.configure.outputs.prev_release }}
+          branch: ${{ needs.configure.outputs.branch }}
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          commit_message: ${{ needs.configure.outputs.commit_message }}
+          pr_list: ${{ needs.prepare.outputs.pr_list }}
+          notes: ${{ needs.prepare.outputs.notes }}
+          dry_run: ${{ inputs.dry_run }}
+          slack_token: ${{ secrets.SLACK_BOT_TOKEN }}
+          slack_channel: ${{ vars.SLACK_SDK_RELEASE_CHANNEL }}
+          label: bt-publishing-test
+          emoji: ':python:'
+
+  # PUBLISH side — gated: verifies the attestation against the downloaded dists, then publishes them.
+  ship-package:
+    needs: [configure, prepare, pack, request-approval]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 15
+    environment: ${{ inputs.dry_run && 'publish-dry-run' || 'publish' }}
+    permissions:
+      contents: write
+      id-token: write     # OIDC trusted publishing + PEP 740
+      attestations: read  # gh attestation verify
+    steps:
+      - name: Download artifact
+        uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8.0.1
+        with:
+          name: release-package
+          path: dl
+
+      - name: Ship package
+        uses: braintrustdata/sdk-actions/actions/release/lang/py/ship-package@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          working_directory: dl   # holds dist/ (whl + sdist) + sbom.json
+          sha: ${{ inputs.sha }}
+          dry_run: ${{ inputs.dry_run }}
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          github_release: ${{ needs.configure.outputs.github_release }}
+          version: ${{ needs.configure.outputs.version }}
+          package_name: bt-publishing-test
+          label: bt-publishing-test
+          notes: ${{ needs.prepare.outputs.notes }}
+          prev_release: ${{ needs.configure.outputs.prev_release }}
+          branch: ${{ needs.configure.outputs.branch }}
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          pr_list: ${{ needs.prepare.outputs.pr_list }}
+          slack_token: ${{ secrets.SLACK_BOT_TOKEN }}
+          slack_channel: ${{ vars.SLACK_SDK_RELEASE_CHANNEL }}
+          emoji: ':python:'
