+++ test/release/ruby/.github/workflows/release.yml
@@ -1,0 +1,141 @@
+# GENERATED by sdk-actions `bin/workflow` — update with: bin/workflow update <this file>
+# sdk-actions: {"template":"release/ruby/turnkey","ref":"38dcec863c910f6802c03a94ab9aeb79931a0a3a","version":"1.0.0","params":{"dry_run_environment":"publish-dry-run","emoji":":gem:","gem_name":"bt-publishing-test","publish_environment":"publish","ruby_version":"","slack_channel_var":"SLACK_SDK_RELEASE_CHANNEL","slack_token_secret":"SLACK_BOT_TOKEN","version_file":"","version_module":"BtPublishingTest","workflow_name":"Release","working_directory":"."}}
+
+name: Release
+
+# Releases bt-publishing-test to RubyGems via braintrustdata/sdk-actions. Dispatch after merging
+# the commit that bumps BtPublishingTest::VERSION; the release runs against that commit's SHA.
+on:
+  workflow_dispatch:
+    inputs:
+      release_type:
+        description: "Release type (stable → GitHub release; prerelease → none)"
+        type: choice
+        default: stable
+        options: [stable, prerelease]
+      sha:
+        description: "Commit SHA (of the version bump) to release"
+        required: true
+        type: string
+      prev_release:
+        description: "Release-notes anchor: a tag or SHA (empty → the previous tag)"
+        type: string
+        required: false
+      dry_run:
+        description: "Dry run: build without tagging or publishing"
+        type: boolean
+        default: false
+
+jobs:
+  configure:
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions:
+      contents: read
+    outputs:
+      version: ${{ steps.configure.outputs.version }}
+      release_tag: ${{ steps.configure.outputs.release_tag }}
+      prev_release: ${{ steps.configure.outputs.prev_release }}
+      branch: ${{ steps.configure.outputs.branch }}
+      on_release_branch: ${{ steps.configure.outputs.on_release_branch }}
+      commit_message: ${{ steps.configure.outputs.commit_message }}
+      github_release: ${{ steps.configure.outputs.github_release }}
+    steps:
+      - name: Configure release
+        id: configure
+        uses: braintrustdata/sdk-actions/actions/release/lang/ruby/configure@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          version_file: lib/bt_publishing_test/version.rb
+          version_module: BtPublishingTest
+          sha: ${{ inputs.sha }}
+          working_directory: .
+          release_type: ${{ inputs.release_type }}
+
+  prepare:
+    needs: configure
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions:
+      contents: write  # releases/generate-notes API
+    outputs:
+      pr_list: ${{ steps.prepare.outputs.pr_list }}
+      notes: ${{ steps.prepare.outputs.notes }}
+    steps:
+      - name: Prepare release
+        id: prepare
+        uses: braintrustdata/sdk-actions/actions/release/prepare@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          sha: ${{ inputs.sha }}
+          prev_release: ${{ inputs.prev_release || needs.configure.outputs.prev_release }}
+
+  validate:
+    needs: [configure, prepare]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 10
+    permissions:
+      contents: read
+    steps:
+      - name: Validate release
+        uses: braintrustdata/sdk-actions/actions/release/lang/ruby/validate@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          version: ${{ needs.configure.outputs.version }}
+          sha: ${{ inputs.sha }}
+          dry_run: ${{ inputs.dry_run }}
+          working_directory: .
+          package_name: bt-publishing-test
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          notes: ${{ needs.prepare.outputs.notes }}
+
+  request-approval:
+    needs: [configure, prepare, validate]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 5
+    permissions: {}
+    steps:
+      - name: Request release approval
+        uses: braintrustdata/sdk-actions/actions/release/request-approval@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          sha: ${{ inputs.sha }}
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          prev_release: ${{ needs.configure.outputs.prev_release }}
+          branch: ${{ needs.configure.outputs.branch }}
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          commit_message: ${{ needs.configure.outputs.commit_message }}
+          pr_list: ${{ needs.prepare.outputs.pr_list }}
+          notes: ${{ needs.prepare.outputs.notes }}
+          dry_run: ${{ inputs.dry_run }}
+          slack_token: ${{ secrets.SLACK_BOT_TOKEN }}
+          slack_channel: ${{ vars.SLACK_SDK_RELEASE_CHANNEL }}
+          label: bt-publishing-test
+          emoji: ':gem:'
+
+  build-and-ship:
+    needs: [configure, prepare, validate, request-approval]
+    runs-on: ubuntu-24.04
+    timeout-minutes: 15
+    environment: ${{ inputs.dry_run && 'publish-dry-run' || 'publish' }}
+    permissions:
+      contents: write
+      id-token: write
+      attestations: write  # signed SBOM + SLSA build provenance
+    steps:
+      - name: Build and ship
+        uses: braintrustdata/sdk-actions/actions/release/lang/ruby/build-and-ship@38dcec863c910f6802c03a94ab9aeb79931a0a3a
+        with:
+          sha: ${{ inputs.sha }}
+          working_directory: .
+          dry_run: ${{ inputs.dry_run }}
+          release_tag: ${{ needs.configure.outputs.release_tag }}
+          github_release: ${{ needs.configure.outputs.github_release }}
+          notes: ${{ needs.prepare.outputs.notes }}
+          prev_release: ${{ needs.configure.outputs.prev_release }}
+          branch: ${{ needs.configure.outputs.branch }}
+          on_release_branch: ${{ needs.configure.outputs.on_release_branch }}
+          pr_list: ${{ needs.prepare.outputs.pr_list }}
+          slack_token: ${{ secrets.SLACK_BOT_TOKEN }}
+          slack_channel: ${{ vars.SLACK_SDK_RELEASE_CHANNEL }}
+          gem_name: bt-publishing-test
+          label: bt-publishing-test
+          emoji: ':gem:'
